Misc. changes to vim config
* Don't disable search highlighting until requested--I'm making more
use of the highlighting now.
* Simplify autocompletion to `Ack!` in command mode--only autocomplete
`ag`. I never use ack.vim without the `!` and ag is the easiest option
to type. Including completion for `ack` doesn't let me search for it in
the help file.
* Add `testing/coverage` to gitignore so it doesn't appear in my ctrlp
or ack.vim results.
* Change 'kill all tabs' tmux command to ctrl-s K. ctrl-s ctrl-k was too
close to other commands I use, and I was killing lots of panes by
accident.