Update agignore and gitignore_global
If I need to remember to include a file or directory in a project's
gitignore file (e.g. testing/coverage), but don't ever want to see it in
my search results, it belongs in my agignore, not my gitignore_global.

Add package-lock.json to agignore--it never includes relevant results in
my searches.

*swp was in both files and belongs in gitignore.