Tidy aliases; improve fzf/ctrlsf configs in vim
Source zsh aliases in .zshenv so that they are accessible in vim. This
means that `ag` will respect ~/.ignore by default whenever it is run in
vim, drying up the config (no need to respecify the alias whenever I
want the custom behaviour in my vim config), and greatly improving the
speed of ctrlsf (as there is no way of adding options to the ag
query inside vim).