Correct script for rebasing a branch from its original base to master
The script now takes one arg, the branch that was its original base, and
uses this to calculate all commits since branching from the original
base.

This is useful when the original base has been merged into master and
you want to tidy up the history of the new branch.