Update beets config
Don't try to fancy-automate various artists modifications as it seems to
mess things up during import.

It's possible the scripting updates were run after my manual changes
which is why the 'song (artist)' transformation looks broken due to
happening 'twice'. But I'm still reverting for now as beets didn't seem
wise to the change in its own db, which is problematic. I might
experiment with this again in future and see if I can get it automatic +
showing up in beets.