Simplify clojure mappings
regular-people forces its bindings, and the only ones that can't be
manually set are the custom surround bindings, so just remove it.

kovisoft/paredit conflicts with vim-sexp a bunch of the time, and
comment-autobalancing is now handled without it, so remove it.

flow bindings for vim-sexp feel more natural, so make them the
'defaults'.

re-enable autopairs in vim-sexp now kovisoft/paredit isn't doing it.

n.b. auto-balance on delete from kovisoft/paredit is now gone.